Creation of the gallery with images + styling of the navigation bar and gallery
I have started the gallery with the addition of styling it and the navigation bar.

I have only added one image so far just to get the hang of it.

I have also added the media code to make it more responsive to other devices as well as some comments for easier navigation