Creation of 2nd Image, background image changing + adjusting and overall changes to the website
Very big Commit:

Addition of the 2nd gallery image, being a wolf.

Addition of the H1 heading for the title of the Gallery

Used more Javascript to change the size and resolution of the background image.

Added specific sources + more commenting where it was needed

Added the FontAwesome script to the website, which allowed me to add icons to the Minigame and Gallery button on the Navigation Bar

The text + images have been centered and have been increased in size to have less space.